src/main/java/io/codelex/dateandtime/practice/DateTimeExercise.java
package io.codelex.dateandtime.practice;

import io.codelex.enums.examples.Day;

import java.time.DayOfWeek;
import java.time.LocalDate;
import java.time.temporal.TemporalAdjuster;
import java.time.temporal.TemporalAdjusters;

public class DateTimeExercise {
    public static final long DAYS_BETWEEN = 5;

    public static LocalDate createNewYearsEve2017() {
        return LocalDate.of(2017, 12, 31);
    }

    public static LocalDate[] getTwoLocalDates() {
        LocalDate[] dates = new LocalDate[2];
        dates[0] = LocalDate.now();
        dates[1] = dates[0].plusDays(DAYS_BETWEEN);
        return dates;
    }

    public static LocalDate findNextFriday13th(LocalDate from) {
        LocalDate result = from.withDayOfMonth(13);
        while (result.getDayOfWeek() != DayOfWeek.FRIDAY) {
            result.plusMonths(1);
        }
        return result;
    }
}
